Consolidate to single recommender script
- Rename emr_recommender_local_s3.py → emr_recommender.py (primary script)
- Rename old emr_recommender.py → emr_recommender_legacy.py (backup)
- Deprecate emr_recommender_dual_mode.py (redirect to new script)
- Update README to use emr_recommender.py as single source of truth
- Single script now supports: local/S3 paths, dual-mode, all features
