+++ utilities/EMR-Serverless-Config-Advisor/emr_recommender_dual_mode.py
@@ -1,14 +1,31 @@
 #!/usr/bin/env python3
 """
-EMR Serverless Recommender - Dual Mode (Cost + Performance)
-Generates two sets of recommendations from the same metrics.
+DEPRECATED: Use emr_recommender.py instead
+
+This file is kept for backward compatibility only.
+The new emr_recommender.py supports:
+- Both local and S3 paths
+- Dual-mode (cost + performance) recommendations
+- All features from this script
+
+Migration:
+  OLD: python3 emr_recommender_dual_mode.py --s3-path s3://bucket/prefix/
+  NEW: python3 emr_recommender.py --input-path s3://bucket/prefix/
 """
 
-import json
-from pathlib import Path
+import sys
+print("=" * 80)
+print("DEPRECATED: emr_recommender_dual_mode.py")
+print("=" * 80)
+print("\nPlease use emr_recommender.py instead:")
+print("  python3 emr_recommender.py --input-path <s3://bucket/prefix or /local/path>")
+print("\nThe new script supports both local and S3 paths with dual-mode optimization.")
+print("=" * 80)
+sys.exit(1)
 
-# Import from the local module
-from emr_recommender import (
+
+# Original code below (kept for reference)
+"""
     EmrRecommender, _select_worker_type, _calculate_shuffle_ratio,
     _compute_exec_limits, _auto_tune_shuffles, _calculate_executor_disk,
     _max_partition_bytes, _get_timeout_configs, _get_s3_retry_configs,
